feat(ui): separate InfluxDB roles page
